pyproject.toml
# Copyright (c) 2025 Mimer Information Technology

# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in all
# copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.
#
# See license for more details.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mimer-mcp-server"
dynamic = ["version"]
description = "A MCP server with Mimer Database Connectivity"
readme = "README.md"
authors = [
    { name = "Chananya Pomkaew", email = "chananya.pomkaew@mimer.com" },
    { name = "Fredrik Ålund", email = "fredrik.alund@mimer.com" },
    { name = "Mimer Information Technology AB", email = "info@mimer.com" }
]
license = "MIT"
license-files = [
    "LICENSE.md",
]
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.13.0",
    "mcp>=1.15.0",
    "mimerpy>=1.3.0",
    "python-dotenv>=1.1.1",
]

keywords = ["mimer", "mimer sql", "database", "mimer mcp server", "mcp", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10"
]

[project.scripts]
mimer_mcp_server = "mimer_mcp_server.server:main"

[project.urls]
Homepage = "https://github.com/mimersql/mimer-mcp-server"
Repository = "https://github.com/mimersql/mimer-mcp-server"
Documentation = "https://github.com/mimersql/mimer-mcp-server#readme"
Issues = "https://github.com/mimersql/mimer-mcp-server/issues"

[tool.hatch.version]
path = "src/mimer_mcp_server/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/mimer_mcp_server"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "mkdocs-exclude>=1.0.2",
    "mkdocs-material>=9.7.0",
    "mkdocs-snippets>=1.3.2",
    "mkdocstrings[python]>=0.30.1",
    "ipykernel>=7.1.0",
]
